[CHORE #160409504]- Move answer classes
